shell: Allow more parallelism in the playbook
In this playbook, we only care about the order of actions on each
individual server: it's safe to let servers run through the playbook
without waiting on other servers.

This leaves synchronisation between the two plays (pulling in changes
from git and applying `apt` changes) which is unnecessary.
On the other hand, there is very little cost to it anyhow, as most of
the execution time is spent on the `apt` play.